// src/types/distributor.ts
export interface Distributor {
  id: string; // Unique identifier for the distributor
  firstName: string; // First name of the distributor
  lastName: string; // Last name of the distributor
  phone: string; // Phone number of the distributor
  isActive: boolean; // Status of the distributor (Active/Inactive)
  userName: string; // New field for the username
  orderCount: number; // New field for the username

}